Add Outlet and NavLink components
