+++ src/components/CalculatorComponent.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import './calc.css';
+import '../styles/calc.css';
 import CalcButton from './CalcButton';
 import CalcDisplay from './CalcDisplay';
 import calculate from '../logic/calculate';
