+++ src/components/Navbar.jsx
@@ -1,16 +1,24 @@
+import { NavLink } from 'react-router-dom';
+
 const Navbar = () => {
+  const links = [
+    { path: '/', text: 'Home' },
+    { path: 'calculator', text: 'Calculator' },
+    { path: 'quote', text: 'Quote' },
+  ];
   return (
     <nav>
       <h1>Math Magicians</h1>
       <ul>
-        <li>Home</li>
-        <li>Calculator</li>
-        <li>Quote</li>
+        { links.map((link) => (
+          <li key={link.text}>
+            <NavLink to={link.path}>{link.text}</NavLink>
+          </li>
+        ))}
       </ul>
 
-    </nav>    
+    </nav>
   );
-
 };
 
 export default Navbar;
